# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Vyakti Parichay API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface SocialLinks {
  /** @nullable */
  linkedin?: string | null;
  /** @nullable */
  twitter?: string | null;
  /** @nullable */
  instagram?: string | null;
  /** @nullable */
  facebook?: string | null;
  /** @nullable */
  youtube?: string | null;
}

export interface Profile {
  id: number;
  name: string;
  category: string;
  tagline: string;
  /** @nullable */
  bio?: string | null;
  /** @nullable */
  location?: string | null;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  website?: string | null;
  /** @nullable */
  avatarUrl?: string | null;
  /** @nullable */
  coverImageUrl?: string | null;
  achievements?: string[];
  awards?: string[];
  galleryUrls?: string[];
  socialLinks?: SocialLinks;
  featured?: boolean;
  verified?: boolean;
  /** @nullable */
  yearsOfExperience?: number | null;
  /** @nullable */
  organization?: string | null;
  createdAt: string;
  /** @nullable */
  updatedAt?: string | null;
}

export interface ProfileInput {
  /** @minLength 2 */
  name: string;
  category: string;
  /** @minLength 5 */
  tagline: string;
  bio?: string;
  location?: string;
  phone?: string;
  email?: string;
  website?: string;
  avatarUrl?: string;
  coverImageUrl?: string;
  achievements?: string[];
  awards?: string[];
  galleryUrls?: string[];
  socialLinks?: SocialLinks;
  yearsOfExperience?: number;
  organization?: string;
}

export interface ProfileUpdate {
  name?: string;
  category?: string;
  tagline?: string;
  bio?: string;
  location?: string;
  phone?: string;
  email?: string;
  website?: string;
  avatarUrl?: string;
  coverImageUrl?: string;
  achievements?: string[];
  awards?: string[];
  galleryUrls?: string[];
  socialLinks?: SocialLinks;
  featured?: boolean;
  verified?: boolean;
  yearsOfExperience?: number;
  organization?: string;
}

export interface ProfileList {
  profiles: Profile[];
  total: number;
  page: number;
  limit: number;
}

export interface CategoryCount {
  category: string;
  count: number;
}

export interface PlatformStats {
  totalProfiles: number;
  featuredProfiles: number;
  categoriesCount: number;
  verifiedProfiles: number;
  categoryBreakdown?: CategoryCount[];
}

export interface ContactInput {
  /** @minLength 2 */
  name: string;
  email: string;
  phone?: string;
  subject?: string;
  /** @minLength 10 */
  message: string;
}

export interface ContactMessage {
  id: number;
  name: string;
  email: string;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  subject?: string | null;
  message: string;
  createdAt: string;
}

export type ListProfilesParams = {
category?: string;
search?: string;
featured?: boolean;
page?: number;
limit?: number;
};

export type GetFeaturedProfilesParams = {
limit?: number;
};