# lib/api-zod/src/generated/types/contactInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Vyakti Parichay API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ContactInput {
  /** @minLength 2 */
  name: string;
  email: string;
  phone?: string;
  subject?: string;
  /** @minLength 10 */
  message: string;
}