# lib/api-zod/src/generated/types/contactMessage.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Vyakti Parichay API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ContactMessage {
  id: number;
  name: string;
  email: string;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  subject?: string | null;
  message: string;
  createdAt: Date;
}