# lib/api-zod/src/generated/types/platformStats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Vyakti Parichay API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CategoryCount } from './categoryCount';

export interface PlatformStats {
  totalProfiles: number;
  featuredProfiles: number;
  categoriesCount: number;
  verifiedProfiles: number;
  categoryBreakdown?: CategoryCount[];
}