# lib/api-zod/src/generated/types/profile.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Vyakti Parichay API specification
 * OpenAPI spec version: 0.1.0
 */
import type { SocialLinks } from './socialLinks';

export interface Profile {
  id: number;
  name: string;
  category: string;
  tagline: string;
  /** @nullable */
  bio?: string | null;
  /** @nullable */
  location?: string | null;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  website?: string | null;
  /** @nullable */
  avatarUrl?: string | null;
  /** @nullable */
  coverImageUrl?: string | null;
  achievements?: string[];
  awards?: string[];
  galleryUrls?: string[];
  socialLinks?: SocialLinks;
  featured?: boolean;
  verified?: boolean;
  /** @nullable */
  yearsOfExperience?: number | null;
  /** @nullable */
  organization?: string | null;
  createdAt: Date;
  /** @nullable */
  updatedAt?: Date | null;
}