# lib/api-zod/src/generated/types/profileInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Vyakti Parichay API specification
 * OpenAPI spec version: 0.1.0
 */
import type { SocialLinks } from './socialLinks';

export interface ProfileInput {
  /** @minLength 2 */
  name: string;
  category: string;
  /** @minLength 5 */
  tagline: string;
  bio?: string;
  location?: string;
  phone?: string;
  email?: string;
  website?: string;
  avatarUrl?: string;
  coverImageUrl?: string;
  achievements?: string[];
  awards?: string[];
  galleryUrls?: string[];
  socialLinks?: SocialLinks;
  yearsOfExperience?: number;
  organization?: string;
}