# lib/api-zod/src/generated/types/profileList.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Vyakti Parichay API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Profile } from './profile';

export interface ProfileList {
  profiles: Profile[];
  total: number;
  page: number;
  limit: number;
}