# lib/api-zod/src/generated/types/profileUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Vyakti Parichay API specification
 * OpenAPI spec version: 0.1.0
 */
import type { SocialLinks } from './socialLinks';

export interface ProfileUpdate {
  name?: string;
  category?: string;
  tagline?: string;
  bio?: string;
  location?: string;
  phone?: string;
  email?: string;
  website?: string;
  avatarUrl?: string;
  coverImageUrl?: string;
  achievements?: string[];
  awards?: string[];
  galleryUrls?: string[];
  socialLinks?: SocialLinks;
  featured?: boolean;
  verified?: boolean;
  yearsOfExperience?: number;
  organization?: string;
}